Move ConfigurationStrategy.Configuration to top level.
