+++ pluginapi/src/main/java/org/robolectric/pluginapi/SdkPicker.java
@@ -2,7 +2,7 @@ package org.robolectric.pluginapi;
 
 import java.util.List;
 import javax.annotation.Nonnull;
-import org.robolectric.pluginapi.config.ConfigurationStrategy.Configuration;
+import org.robolectric.pluginapi.config.Configuration;
 
 public interface SdkPicker {
