+++ pluginapi/src/main/java/org/robolectric/pluginapi/config/ConfigurationStrategy.java
@@ -1,8 +1,6 @@
 package org.robolectric.pluginapi.config;
 
 import java.lang.reflect.Method;
-import java.util.Collection;
-import java.util.Map;
 
 /**
  * Strategy for configuring individual tests.
@@ -22,22 +20,10 @@ public interface ConfigurationStrategy {
    * @param method the test method to be evaluated
    * @return the set of configs
    */
-  Configuration getConfig(Class<?> testClass, Method method);
+  org.robolectric.pluginapi.config.Configuration getConfig(Class<?> testClass, Method method);
 
-  /**
-   * Heterogeneous typesafe collection of configuration objects managed by their {@link Configurer}.
-   *
-   * @since 4.2
-   */
-  interface Configuration {
-
-    /** Returns the configuration instance of the specified class for the current test. */
-    <T> T get(Class<T> configClass);
-
-    /** Returns the set of known configuration classes. */
-    Collection<Class<?>> keySet();
-
-    /** Returns the map of known configuration classes to configuration instances. */
-    Map<Class<?>, Object> map();
+  /** @deprecated Use {@link org.robolectric.pluginapi.config.Configuration} instead. */
+  @Deprecated
+  interface Configuration extends org.robolectric.pluginapi.config.Configuration {
   }
 }
