+++ robolectric/src/main/java/org/robolectric/plugins/DefaultSdkPicker.java
@@ -21,7 +21,7 @@ import org.robolectric.annotation.internal.ConfigUtils;
 import org.robolectric.pluginapi.Sdk;
 import org.robolectric.pluginapi.SdkPicker;
 import org.robolectric.pluginapi.UsesSdk;
-import org.robolectric.pluginapi.config.ConfigurationStrategy.Configuration;
+import org.robolectric.pluginapi.config.Configuration;
 
 /** Robolectric's default {@link SdkPicker}. */
 @SuppressWarnings("NewApi")
