+++ robolectric/src/test/java/org/robolectric/SingleSdkRobolectricTestRunner.java
@@ -8,7 +8,7 @@ import org.junit.runners.model.InitializationError;
 import org.robolectric.pluginapi.Sdk;
 import org.robolectric.pluginapi.SdkPicker;
 import org.robolectric.pluginapi.UsesSdk;
-import org.robolectric.pluginapi.config.ConfigurationStrategy.Configuration;
+import org.robolectric.pluginapi.config.Configuration;
 import org.robolectric.util.TestUtil;
 import org.robolectric.util.inject.Injector;
